Improve UI | Add Font and Assets | Fix catch no connectivity error

+++ lib/model/city_suggestion_weather_model.dart
@@ -102,10 +102,10 @@ class Main {
 
   factory Main.fromJson(Map<String, dynamic> json) {
     return Main(
-      temp: json['temp'],
-      feelsLike: json['feels_like'],
-      tempMin: json['temp_min'],
-      tempMax: json['temp_max'],
+      temp: json['temp'].toDouble(),
+      feelsLike: json['feels_like'].toDouble(),
+      tempMin: json['temp_min'].toDouble(),
+      tempMax: json['temp_max'].toDouble(),
       pressure: json['pressure'],
       humidity: json['humidity'],
     );
@@ -120,7 +120,7 @@ class Wind {
 
   factory Wind.fromJson(Map<String, dynamic> json) {
     return Wind(
-      speed: json['speed'],
+      speed: json['speed'].toDouble(),
       deg: json['deg'],
     );
   }
